part1/requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.13.4
cachetools==4.2.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cssselect==1.3.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.1
fsspec==2024.6.1
idna==3.10
jieba3k==0.35.1
Jinja2==3.1.4
joblib==1.5.1
lxml==5.4.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.1.2
nvidia-cublas-cu12==12.8.3.14
nvidia-cuda-cupti-cu12==12.8.57
nvidia-cuda-nvrtc-cu12==12.8.61
nvidia-cuda-runtime-cu12==12.8.57
nvidia-cudnn-cu12==9.7.1.26
nvidia-cufft-cu12==11.3.3.41
nvidia-cufile-cu12==1.13.0.11
nvidia-curand-cu12==10.3.9.55
nvidia-cusolver-cu12==11.7.2.55
nvidia-cusparse-cu12==12.5.7.53
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.8.55
pillow==11.0.0
python-dateutil==2.9.0.post0
python-telegram-bot==13.15
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
setuptools==70.2.0
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.7
sympy==1.13.3
tinysegmenter==0.3
tldextract==5.3.0
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchvision==0.22.1+cu128
tornado==6.1
tqdm==4.67.1
triton==3.3.1
typing_extensions==4.12.2
tzlocal==5.3.1
urllib3==2.4.0
